Allow clients to control the Wayland layer shell layer
Notifications that have a "overlay" (boolean) hint set will be placed in the
OVERLAY layer of the Wayland layer shell [1]. Otherwise, they will be placed
by default in the TOP layer. This hint has no impact if we are not
running in Wayland (runtime check against GDK_IS_WAYLAND_DISPLAY).

[1]: https://github.com/swaywm/wlroots/blob/master/protocol/wlr-layer-shell-unstable-v1.xml